+++ ruby-lsp-ree/CHANGELOG.md
@@ -18,3 +18,4 @@
 - autocomplete for ree actions
 - Add Link for ree actions
 - sort links in ree actions
+- autocomplete for ree dao